bcache-tools: Remove the dependency on libsmartcols
The bcache tree function relies on the libsmartcols library to
display the relationship between the cache device and the bdev
device in a tree shape. However, it is difficult for many old
operating systems (such as Ubuntu 12) to install this library.

For better compatibility, a simpler implementation is used to
achieve the same purpose, while removing the dependency on
libsmartcols.

Signed-off-by: Shaoxiong Li <dahefanteng@gmail.com>
Signed-off-by: Coly Li <colyli@suse.de>